chore(server-split): v0.4.x PR-D — W7-A artifacts + upload-history routes → routes/artifacts.py (5 endpoints)
docs/design/v0.4.x-server-split.md 8-PR sequence 다섯 번째 PR.
PR-C 위에 W7-A 데이터 아티팩트 + upload history 분리.

## 이동된 surface (158 lines 서버에서 제거)

### 5 endpoints
- /admin/uploads/history/  (74L)
- /admin/artifacts/list, /admin/artifacts/{artifact_id}
- /artifacts/mine/list, /artifacts/mine/{artifact_id}

Pydantic 모델 없음. SQL audit_log 조회 패턴 — _AUDIT_DB + sqlite3 import.

## server_llmwiki.py 변경
- 5 endpoint 제거 (~158 lines)
- `from routes.artifacts import router as artifacts_router` + `app.include_router(artifacts_router)` (jobs_router 다음)

## tests/_server_split_helpers.py + test_upload_history.py
- helper module list 에 `routes.artifacts` 추가
- test_upload_history.py 의 setUpClass 가 combined_server_source() 사용 (source-level 스캔 6 테스트 — 다른 server-split test 와 동일 패턴)

## 검증 (design memo §5 contract)

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ 125 endpoints identical to origin/main |
| 2-5. middleware/audit/static/startup | ✓ smoke test pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** |
| ruff F-class | ✓ All checks passed |

server_llmwiki.py: 4439 → 4281 lines (-158).
8-PR roadmap 진행률 ≈ **27%** (1613/5894).

## Out of scope (다음 PR)
- PR-E: Phase 7 자기진화 + 멀티모달
- PR-F: /code/*
- PR-G: 잔여 /admin/*
- PR-H: server 최종 정리

Quality delta: exempt (label: chore). URL byte-identical, core/ 무관.
